moved test_legacy_device_ids_are_updated to test_init.py
